Typescriptify Card component
