+++ types/packages.d.ts
@@ -1,1 +1,3 @@
+declare module 'prop-types';
 declare module 'react-bar-chart';
+declare module 'react-textfit';
